move removeComplexExpressions to a midend pass
