+++ backends/dpdk/dpdkHelpers.cpp
@@ -618,40 +618,18 @@ bool ConvertStatementToDpdk::preorder(const IR::MethodCallStatement *s) {
                         append_parser_name(parser, IR::ParserState::reject)));
             add_instr(new IR::DpdkLabelStatement(end_label));
         } else if (a->method->name == "add_entry") {
-            if (a->expr->arguments->size() != 2)
-                ::error("%1%: must have two arguments", a->method);
-
             auto action = a->expr->arguments->at(0)->expression;
-            if (!action->is<IR::StringLiteral>())
-                ::error(ErrorType::ERR_UNEXPECTED, "%1% expected string in first argument", s);
             auto action_name = action->to<IR::StringLiteral>()->value;
-
             auto param = a->expr->arguments->at(1)->expression;
-            IR::Expression* argument = nullptr;
-            if (auto constant = param->to<IR::Constant>()) {
-                if (structure->actions.count(action_name) != 0) {
-                    auto action = structure->actions.at(action_name);
-                    auto params = action->parameters->parameters;
-                    auto field_name = IR::ID(action_name + "_" + params.at(0)->name);
-                    add_instr(new IR::DpdkMovStatement(
-                                new IR::Member(new IR::PathExpression("m"), field_name),
-                                constant));
-                    argument = new IR::Member(new IR::PathExpression("m"), field_name);
-                } else {
-                }
-            } else if (auto params = param->to<IR::StructExpression>()) {
-                for (auto p : params->components) {
-                    auto field_name = IR::ID(action_name + "_" + p->name);
-                    add_instr(new IR::DpdkMovStatement(
-                                new IR::Member(new IR::PathExpression("m"), field_name), new IR::Constant(0)));
-                }
-                auto arg0 = params->components.at(0);
-                auto field0 = IR::ID(action_name + "_" + arg0->name);
-                argument = new IR::Member(new IR::PathExpression("m"), field0);
+            if (param->is<IR::Member>()) {
+                auto argument = param->to<IR::Member>();
+                add_instr(new IR::DpdkLearnStatement(action_name, argument));
+            } else if (param->is<IR::StructExpression>()) {
+                auto argument = param->to<IR::StructExpression>()->components.at(0)->expression;
+                add_instr(new IR::DpdkLearnStatement(action_name, argument));
             } else {
-                ::error(ErrorType::ERR_UNSUPPORTED, "argument %1% is not supported", param);
+                ::error("%1%: unhandled function", s);
             }
-            add_instr(new IR::DpdkLearnStatement(action_name, argument));
         } else if (a->method->name == "send_to_port") {
             add_instr(new IR::DpdkMovStatement(
                 new IR::Member(new IR::PathExpression("m"), "pna_main_output_metadata_output_port"),
